LayoutTests/fast/forms/color/color-input-swatch-expected.txt

initial state

getComputedStyle(swatch).backgroundColor is rgb(0, 0, 0)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: limited-srgb, value: papayawhip

getComputedStyle(swatch).backgroundColor is rgb(255, 239, 213)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: limited-srgb, value: papayawhip

getComputedStyle(swatch).backgroundColor is color(srgb 1 0.937255 0.835294)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: display-p3, value: papayawhip

getComputedStyle(swatch).backgroundColor is color(display-p3 0.989244 0.939427 0.846367)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: display-p3, value: papayawhip

getComputedStyle(swatch).backgroundColor is color(display-p3 0.989244 0.939427 0.846367)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: limited-srgb, value: #44444444

getComputedStyle(swatch).backgroundColor is rgb(68, 68, 68)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: limited-srgb, value: #44444444

getComputedStyle(swatch).backgroundColor is rgb(205, 205, 205)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: display-p3, value: #44444444

getComputedStyle(swatch).backgroundColor is color(display-p3 0.266667 0.266667 0.266667)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: display-p3, value: #44444444

getComputedStyle(swatch).backgroundColor is rgb(205, 205, 205)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: limited-srgb, value: color(display-p3 2 none .5 / .7)

getComputedStyle(swatch).backgroundColor is rgb(255, 0, 104)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: limited-srgb, value: color(display-p3 2 none .5 / .7)

getComputedStyle(swatch).backgroundColor is rgb(255, 76, 149)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: display-p3, value: color(display-p3 2 none .5 / .7)

getComputedStyle(swatch).backgroundColor is color(display-p3 2 0 0.5)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: display-p3, value: color(display-p3 2 none .5 / .7)

getComputedStyle(swatch).backgroundColor is rgb(255, 76, 149)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: limited-srgb, value: rgba(40,40,40,.6)

getComputedStyle(swatch).backgroundColor is rgb(40, 40, 40)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: limited-srgb, value: rgba(40,40,40,.6)

getComputedStyle(swatch).backgroundColor is rgb(126, 126, 126)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: false, colorSpace: display-p3, value: rgba(40,40,40,.6)

getComputedStyle(swatch).backgroundColor is color(display-p3 0.156863 0.156863 0.156863)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto

alpha: true, colorSpace: display-p3, value: rgba(40,40,40,.6)

getComputedStyle(swatch).backgroundColor is rgb(126, 126, 126)
getComputedStyle(swatch).backgroundImage is none
getComputedStyle(swatch).backgroundSize is auto
